+++ BeeOverlay/PluginController.cs
@@ -3,6 +3,7 @@ using BeeOverlay.Core.Ports;
 using BeeOverlay.Core.UseCases;
 using BeeOverlay.Interop;
 using BeeOverlay.Interop.Game;
+using BepInEx.Configuration;
 using BepInEx.Logging;
 
 namespace BeeOverlay;
@@ -13,14 +14,17 @@ namespace BeeOverlay;
 internal sealed class PluginController
 {
     private readonly FrameHandler frameHandler;
+    private readonly BeeOverlayConfiguration configuration;
 
-    private PluginController(FrameHandler frameHandler)
+    private PluginController(FrameHandler frameHandler, BeeOverlayConfiguration configuration)
     {
         this.frameHandler = frameHandler;
+        this.configuration = configuration;
     }
 
-    public static PluginController Create(ManualLogSource logger)
+    public static PluginController Create(ManualLogSource logger, ConfigFile config)
     {
+        BeeOverlayConfiguration configuration = BeeOverlayConfiguration.Bind(config);
         IOverlayObservationSource observationSource = new BeeObservationSource(logger);
         IOverlayPresenter presenter = new Overlay(logger);
         var frameHandler = new FrameHandler(
@@ -29,11 +33,11 @@ internal sealed class PluginController
             buildOverlayFrameUseCase: new BuildOverlayFrameUseCase()
         );
 
-        return new PluginController(frameHandler);
+        return new PluginController(frameHandler, configuration);
     }
 
     public void HandleFrame()
     {
-        frameHandler.HandleFrame();
+        frameHandler.HandleFrame(configuration.Enabled, configuration.PresentationOptions);
     }
 }